requirements.txt
Flask==1.0.2
gunicorn==19.9.0
google-cloud-storage==1.10.0
numpy==1.15.0
scipy==1.1.0
sklearn==0.19.2
